Change datatype for scores and some other fields; Implement other criteria values in round robin matches

+++ tournament-app-server/DTOs/MatchSeEditMatchScoreDTO.cs
@@ -3,9 +3,9 @@
     public class MatchSeEditMatchScoreDTO
     {
         public string? winner { get; set; }
-        public long[]? team_1_scores { get; set; }
-        public long[]? team_2_scores { get; set; }
-        public long[]? team_1_subscores { get; set; }
-        public long[]? team_2_subscores { get; set; }
+        public decimal[]? team_1_scores { get; set; }
+        public decimal[]? team_2_scores { get; set; }
+        public decimal[]? team_1_subscores { get; set; }
+        public decimal[]? team_2_subscores { get; set; }
     }
 }
